+++ VendixPos/Models/Inventory.cs
@@ -8,8 +8,8 @@ namespace VendixPos.Models
         public int InventoryID { get; set; }
         public int InvoiceInfoId { get; set; }
         public int InventoryAddID { get; set; }
-        public int InventoryState { get; set; }
-        public decimal InventoryInvSUQ { get; set; }
+        public byte InventoryState { get; set; }
+        public int InventoryInvSUQ { get; set; }
         public int InventoryItemNum { get; set; }
         public int InventoryInvoNum { get; set; }
         public int SupplierID { get; set; }
